requirements.txt
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
astroid==3.3.5
asttokens==2.4.1
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
backoff==2.2.1
billiard==4.2.1
black==24.10.0
blinker==1.9.0
boto3==1.35.36
botocore==1.35.36
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.43
dagshub-annotation-converter==0.1.2
databricks-sdk==0.37.0
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
Deprecated==1.2.14
dictdiffer==0.9.0
-e git+ssh://git@github.com/GianmarcoTurchiano/Digits-Recognition.git@9eb256682c37d2f9ec0fdce83c4d65ce4b9ca291#egg=digits_recognition
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.5
dvc==3.56.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.1.0
fastapi==0.115.5
filelock==3.16.1
flake8==7.1.1
Flask==3.1.0
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
gql==3.5.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gto==1.7.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isort==5.13.2
iterative-telemetry==0.0.9
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kombu==5.4.2
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==3.11.0
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyflakes==3.2.0
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pylint==3.3.1
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3fs==2024.10.0
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.8
semver==3.0.2
setuptools==75.5.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
sqltrie==0.11.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1+cu118
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
treelib==1.7.0
triton==3.1.0
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.16.0
yarl==1.17.1
zc.lockfile==3.0.post1
zipp==3.21.0
